Non-breaking, minor changes
